Add health check endpoint and refactor request handling logic
