+++ pkg/api/svc.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log/slog"
 	"net"
+	"net/http"
 
 	"github.com/ksysoev/deriv-api-bff/pkg/core/request"
 	"github.com/ksysoev/deriv-api-bff/pkg/middleware"
@@ -64,6 +65,7 @@ func NewSevice(cfg *Config, handler BFFService) *Service {
 
 	s.server = server.NewServer(cfg.Listen)
 	s.server.AddChannel(endpoint)
+	s.server.AddHandler("/livez", http.HandlerFunc(s.HealthCheck))
 
 	return s
 }
@@ -94,27 +96,6 @@ func (s *Service) Run(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-// Handle processes a request received on a connection and routes it based on the request type.
-// It takes conn of type wasabi.Connection and r of type wasabi.Request.
-// It returns an error if the request type is unsupported or if the request type is empty.
-// If the request type is core.TextMessage or core.BinaryMessage, it passes the request through to the handler.
-// For other request types, it processes the request using the handler.
-func (s *Service) Handle(conn wasabi.Connection, r wasabi.Request) error {
-	req, ok := r.(*request.Request)
-	if !ok {
-		return fmt.Errorf("unsupported request type: %T", req)
-	}
-
-	switch req.RoutingKey() {
-	case request.TextMessage, request.BinaryMessage:
-		return s.handler.PassThrough(conn, req)
-	case "":
-		return fmt.Errorf("empty request type: %v", req)
-	default:
-		return s.handler.ProcessRequest(conn, req)
-	}
 }
 
 // parse processes a message received over a Wasabi connection and converts it into a core request.
